challenge-5/azure-function-orchestrator/requirements.txt
azure-functions
azure-cosmos
azure-identity
semantic-kernel
asyncio
typing
datetime
json
logging
